+++ src/Paper.php
@@ -331,6 +331,102 @@ trait Paper
     }
 
     /**
+     * @param  ?scalar  $operator
+     * @param  ?scalar  $value
+     * @return PaperQueryBuilder<static>
+     */
+    public static function whereDate(string $column, mixed $operator, mixed $value = null): PaperQueryBuilder
+    {
+        [$operator, $value] = func_num_args() === 2 ? ['=', $operator] : [$operator, $value];
+
+        return static::query()->whereDate($column, $operator, $value);
+    }
+
+    /**
+     * @param  ?scalar  $operator
+     * @param  ?scalar  $value
+     * @return PaperQueryBuilder<static>
+     */
+    public static function orWhereDate(string $column, mixed $operator, mixed $value = null): PaperQueryBuilder
+    {
+        [$operator, $value] = func_num_args() === 2 ? ['=', $operator] : [$operator, $value];
+
+        return static::query()->orWhereDate($column, $operator, $value);
+    }
+
+    /**
+     * @param  ?scalar  $operator
+     * @param  ?scalar  $value
+     * @return PaperQueryBuilder<static>
+     */
+    public static function whereYear(string $column, mixed $operator, mixed $value = null): PaperQueryBuilder
+    {
+        [$operator, $value] = func_num_args() === 2 ? ['=', $operator] : [$operator, $value];
+
+        return static::query()->whereYear($column, $operator, $value);
+    }
+
+    /**
+     * @param  ?scalar  $operator
+     * @param  ?scalar  $value
+     * @return PaperQueryBuilder<static>
+     */
+    public static function orWhereYear(string $column, mixed $operator, mixed $value = null): PaperQueryBuilder
+    {
+        [$operator, $value] = func_num_args() === 2 ? ['=', $operator] : [$operator, $value];
+
+        return static::query()->orWhereYear($column, $operator, $value);
+    }
+
+    /**
+     * @param  ?scalar  $operator
+     * @param  ?scalar  $value
+     * @return PaperQueryBuilder<static>
+     */
+    public static function whereMonth(string $column, mixed $operator, mixed $value = null): PaperQueryBuilder
+    {
+        [$operator, $value] = func_num_args() === 2 ? ['=', $operator] : [$operator, $value];
+
+        return static::query()->whereMonth($column, $operator, $value);
+    }
+
+    /**
+     * @param  ?scalar  $operator
+     * @param  ?scalar  $value
+     * @return PaperQueryBuilder<static>
+     */
+    public static function orWhereMonth(string $column, mixed $operator, mixed $value = null): PaperQueryBuilder
+    {
+        [$operator, $value] = func_num_args() === 2 ? ['=', $operator] : [$operator, $value];
+
+        return static::query()->orWhereMonth($column, $operator, $value);
+    }
+
+    /**
+     * @param  ?scalar  $operator
+     * @param  ?scalar  $value
+     * @return PaperQueryBuilder<static>
+     */
+    public static function whereDay(string $column, mixed $operator, mixed $value = null): PaperQueryBuilder
+    {
+        [$operator, $value] = func_num_args() === 2 ? ['=', $operator] : [$operator, $value];
+
+        return static::query()->whereDay($column, $operator, $value);
+    }
+
+    /**
+     * @param  ?scalar  $operator
+     * @param  ?scalar  $value
+     * @return PaperQueryBuilder<static>
+     */
+    public static function orWhereDay(string $column, mixed $operator, mixed $value = null): PaperQueryBuilder
+    {
+        [$operator, $value] = func_num_args() === 2 ? ['=', $operator] : [$operator, $value];
+
+        return static::query()->orWhereDay($column, $operator, $value);
+    }
+
+    /**
      * @return PaperQueryBuilder<static>
      */
     public static function orderBy(string $column, string $direction = 'asc'): PaperQueryBuilder